custom-llm-server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==0.28.1
pydantic==2.10.3
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.32.3 